+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.8.7 - Release Date: 17/03/2016 #
+* Added functions.admin.php
+* Added uk_get_wp_admin_color_scheme() function (to get the current active color scheme color codes)
+* Added uk_color_scheme() function (to output custom css to complement with the current active color scheme)
+
 # 0.8.6 - Release Date: 15/03/2016 #
 * Fixed FontAwesome and Socicon to load local instead of CDN (this is currently not possible due to loading mechanism)